TweeterApp-3/app/src/main/java/edu/byu/cs/tweeter/client/model/service/PostService.java
package edu.byu.cs.tweeter.client.model.service;

import java.io.IOException;

import edu.byu.cs.tweeter.client.model.net.ServerFacade;
import edu.byu.cs.tweeter.client.model.net.TweeterRemoteException;
import edu.byu.cs.tweeter.client.model.service.request.PostRequest;
import edu.byu.cs.tweeter.client.model.service.response.PostResponse;

/**
 * Contains the business logic to support the login operation.
 */
public interface PostService {

    public PostResponse post(PostRequest request) throws IOException, TweeterRemoteException;



    /**
     * Returns an instance of {@link ServerFacade}. Allows mocking of the ServerFacade class for
     * testing purposes. All usages of ServerFacade should get their ServerFacade instance from this
     * method to allow for proper mocking.
     *
     * @return the instance.
     */
   public ServerFacade getServerFacade();
}
